# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ZapStore API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: products
    description: Product operations
  - name: tickets
    description: Ticket operations
  - name: updates
    description: News/update operations
  - name: stats
    description: Statistics operations
  - name: upload
    description: File upload operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - in: query
          name: all
          schema:
            type: string
          description: If "true", return all products including inactive (admin)
        - in: query
          name: category
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: List tickets (admin sees all, user sees own)
      responses:
        "200":
          description: List of tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create an order ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"

  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get a ticket by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "404":
          description: Not found
    patch:
      operationId: updateTicket
      tags: [tickets]
      summary: Update ticket status/notes (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketUpdate"
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"

  /updates:
    get:
      operationId: listUpdates
      tags: [updates]
      summary: List all news updates
      responses:
        "200":
          description: List of updates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Update"
    post:
      operationId: createUpdate
      tags: [updates]
      summary: Create a news update (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInput"
      responses:
        "201":
          description: Update created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Update"

  /updates/{id}:
    get:
      operationId: getUpdate
      tags: [updates]
      summary: Get an update by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Update"
        "404":
          description: Not found
    delete:
      operationId: deleteUpdate
      tags: [updates]
      summary: Delete a news update (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Get store statistics
      responses:
        "200":
          description: Store statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreStats"

  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"

  /admin/users/{id}:
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update user role (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
    delete:
      operationId: deleteAdminUser
      tags: [admin]
      summary: Delete a user (owner only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    User:
      type: object
      required: [id, username, email, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string

    AdminUser:
      type: object
      required: [id, username, email, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string

    AdminUserUpdate:
      type: object
      properties:
        role:
          type: string

    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 2
        email:
          type: string
        password:
          type: string
          minLength: 6

    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"

    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    UploadResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string

    Product:
      type: object
      required: [id, name, description, price, category, status, active, featured, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, coming_soon, not_for_sale]
        active:
          type: boolean
        featured:
          type: boolean
        createdAt:
          type: string

    ProductInput:
      type: object
      required: [name, description, price, category]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        price:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, coming_soon, not_for_sale]
        active:
          type: boolean
        featured:
          type: boolean

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: string
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, coming_soon, not_for_sale]
        active:
          type: boolean
        featured:
          type: boolean

    Ticket:
      type: object
      required: [id, userId, productId, status, message, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        productId:
          type: integer
        status:
          type: string
        message:
          type: string
        adminNotes:
          type: ["string", "null"]
        username:
          type: ["string", "null"]
        productName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    TicketInput:
      type: object
      required: [productId, message]
      properties:
        productId:
          type: integer
        message:
          type: string

    TicketUpdate:
      type: object
      properties:
        status:
          type: string
        adminNotes:
          type: string

    Update:
      type: object
      required: [id, title, content, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        authorUsername:
          type: ["string", "null"]
        createdAt:
          type: string

    UpdateInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string

    StoreStats:
      type: object
      required: [totalProducts, openTickets, totalUsers, latestUpdate]
      properties:
        totalProducts:
          type: integer
        openTickets:
          type: integer
        totalUsers:
          type: integer
        latestUpdate:
          type: ["string", "null"]