# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ZapStore API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductStatus } from './productStatus';

export interface Product {
  id: number;
  name: string;
  description: string;
  price: string;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  status: ProductStatus;
  active: boolean;
  featured: boolean;
  createdAt: string;
}